refactor: renames "index.ts" to "definition.ts" in "library/UniformResourceIdentifier/Data/Image"
