+++ src/features/TextToImage/models/SDXL/client/conversions/StabilityAI_Client_Image.ts
@@ -1,7 +1,7 @@
 import type * as StabilityAIClient from 'stabilityai-client-typescript/models/components';
 
 import Base64CharacterEncodedByteSequence from '@/library/Base64CharacterEncodedByteSequence/definition';
-import ImageURI from '@/library/UniformResourceIdentifier/Data/Image';
+import ImageURI from '@/library/UniformResourceIdentifier/Data/Image/definition';
 
 import type {
   Output,
