+++ src/library/UniformResourceIdentifier/Data/Image/definition.ts
